\docType{data}
\name{anket}
\alias{anket}
\title{'General Social Survey' (GSS) araştırmasından kategorik değişken örnekleri}
\format{21.473 satır ve 9 sütundan oluşan bir tibble
\describe{
\item{yıl}{Araştırma yılı, 2000-2014}
\item{medeni_durum}{Medeni durum}
\item{yaş}{Yaş. Maksimum 89 ile sınırlı}
\item{ırk}{Irk}
\item{gelir}{Bildirilen gelir (ABD Doları)}
\item{parti}{Siyasi parti tercihi}
\item{din}{Dini inanç}
\item{mezhep}{Bağlı olduğu mezhep}
\item{tv_saati}{Günlük televizyon izleme süresi}
}}
\usage{anket}
\description{'General Social Survey' (GSS) araştırmasından kategorik değişken örnekleri}
\seealso{\code{\link[forcats]{gss_cat}}}
\keyword{datasets}
