\docType{data}
\name{atmosfer}
\alias{atmosfer}
\title{Atmosferik veriler}
\format{41.472 satır ve 11 sütundan oluşan bir tibble
\describe{
\item{enlem}{Ölçüm noktası. 113,8 - 56,2 Batı boylamları ve 36,2 Kuzey - 21,2 Güney enlemleri arasında, eşit aralıklı 24 x 24 mekansal grid}
\item{boylam}{Ölçüm noktası. 113,8 - 56,2 Batı boylamları ve 36,2 Kuzey - 21,2 Güney enlemleri arasında, eşit aralıklı 24 x 24 mekansal grid}
\item{yıl}{Ocak 1995'ten Aralık 2000'e kadar, birer ay ara ile 72 gözlem}
\item{ay}{Ocak 1995'ten Aralık 2000'e kadar, birer ay ara ile 72 gözlem}
\item{yüzey_sıcaklığı}{Açık gökyüzü koşullarında dünya yüzeyinden yayılan enerjiye dayalı Kelvin cinsinden aylık ortalama yüzey sıcaklığı}
\item{sıcaklık}{Dünya yüzeyine yakın havanın Kelvin cinsinden aylık ortalama sıcaklığı}
\item{basınç}{Dünya yüzeyinde belirli bir konumdaki aylık ortalama atmosferik yüzey basıncı (mbar)}
\item{ozon}{Ortalama Ozon Bolluğu: Atmosferik kolondaki, Dobson cinsinden aylık ortalama toplam ozon miktarı}
\item{alçak_bulut}{Ortalama Alçak Bulut Örtüsü: Bulut tepe basıncı 680 mbar'dan fazla veya kabaca 3,24 km'den az bulutlarla kaplı gökyüzünün aylık ortalama yüzdesi}
\item{orta_bulut}{Ortalama Orta Bulut Örtüsü: Bulut tepe basıncı 440 - 680 mbar arasında veya kabaca 3,24'ten 6,5 km'ye kadar bulutlarla kaplı gökyüzünün aylık ortalama yüzdesi}
\item{yüksek_bulut}{Ortalama Yüksek Bulut Örtüsü: Bulut tepe basıncı 440 mbar'a eşit ya da daha az veya kabaca 6.5 km'den daha fazla bulutlarla kaplı gökyüzünün aylık ortalama yüzdesi}
}}
\usage{atmosfer}
\description{Atmosferik veriler.}
\seealso{\code{\link[nasaweather]{atmos}}}
\keyword{datasets}
