\docType{data}
\name{başkanlık}
\alias{başkanlık}
\title{Eisenhower'dan Obama'ya 11 başkanın dönemleri}
\format{11 satır ve 4 sütundan oluşan bir data.frame
\describe{
\item{ad}{Başkanın soyadı}
\item{başlangıç}{Başkanlığın başlangıç tarihi}
\item{bitiş}{Başkanlığın bitiş tarihi}
\item{parti}{Üyesi olduğu siyasi parti}
}}
\usage{başkanlık}
\description{Eisenhower'dan Obama'ya kadar 11 ABD başkanının ve siyasi partisinin adı, başkanlık dönemlerimin başlangıç ve bitiş tarihi.}
\seealso{\code{\link[ggplot2]{presidential}}}
\keyword{datasets}
