\docType{data}
\name{bebek_isimleri}
\alias{bebek_isimleri}
\title{Bebek isimleri}
\format{1.924.665 satır ve 5 sütundan oluşan bir tibble
\describe{
\item{yıl}{Doğum yılı}
\item{cinsiyet}{Bebeğin cinsiyeti}
\item{isim}{İsim}
\item{sayı}{Bebek sayısı}
\item{oran}{Belirtilen cinsiyet ve yıl için aynı isme sahip olanların oranı}
}}
\usage{bebek_isimleri}
\description{ABD Sosyal Güvenlik Kurumu'ndan elde edilen bebek ismi verileri. En az 5 kullanıma sahip tüm isimleri içerir.}
\seealso{\code{\link[babynames]{babynames}}}
\keyword{datasets}
