\docType{data}
\name{elmaslar}
\alias{elmaslar}
\title{50 bin adetten fazla yuvarlak kesim elmasın fiyatları}
\format{53.940 gözlem ve 10 değişkenden oluşan bir data.frame
\describe{
\item{karat}{Elmasın ağırlığı (0,2-5,01 ct)}
\item{kesim}{Kesim kalitesi (Vasat, İyi, Çok iyi, Birinci sınıf, İdeal)}
\item{renk}{Elmas rengi, D'den (en iyi) J'ye (en kötü)}
\item{berraklık}{Elmasın berraklık ölçüsü (I1 (en kötü), SI1, SI2, VS1, VS2, VVS1, VVS2, IF (en iyi))}
\item{derinlik}{Toplam derinlik yüzdesi = z / mean(x, y) = 2 * z / (x + y) (43-79)}
\item{yüzey}{Elmasın düz üst yüzeyinin genişliği (43-95 mm)}
\item{fiyat}{Amerikan doları cinsinden fiyat (326$-18.823$)}
\item{x}{Uzunluk (mm)}
\item{y}{Genişlik (mm)}
\item{z}{Derinlik (mm)}
}}
\usage{elmaslar}
\description{Neredeyse 54 bin elmasın fiyatlarını ve diğer özelliklerini içeren bir veri kümesi.}
\seealso{\code{\link[ggplot2]{diamonds}}}
\keyword{datasets}
