\docType{data}
\name{gapminder_tr}
\alias{gapminder_tr}
\title{Gapminder verileri.}
\format{1.704 satır ve 6 sütundan oluşan bir tibble
\describe{
\item{ülke}{Ülke adları (142 faktör)}
\item{kıta}{Kıtalar (5 faktör)}
\item{yıl}{1952'den 2007'ye kadar 5 yıllık artışlarla}
\item{ort_ömür}{Beklenen yaşam süresi (yıl)}
\item{nüfus}{Nüfus}
\item{gsyih}{Kişi başına düşen gayri safi yurt içi hasıla (enflasyona göre düzeltilmiş ABD Doları cinsinden)}
}}
\usage{gapminder_tr}
\description{Ülkelere göre beklenen yaşam süresi, kişi başına GSYİH ve nüfus hakkındaki Gapminder verileri.}
\source{http://www.gapminder.org/data/}
\seealso{\code{\link[gapminder]{gapminder}}}
\keyword{datasets}
