\docType{data}
\name{havalimanları}
\alias{havalimanları}
\title{Havalimanları kılavuzu}
\format{1.458 satır ve 8 sütundan oluşan bir data.frame
\describe{
\item{havalimanı_kodu}{FAA havalimanı kodu}
\item{ad}{Havalimanı adı}
\item{enlem}{Havalimanının konumu}
\item{boylam}{Havalimanının konumu}
\item{irtifa}{İrtifa (ft)}
\item{saat_dilimi}{Greenwich'e göre saat dilimi farkı}
\item{ysu}{Yaz saati uygulaması, Gün işiğindan daha fazla yararlanmak ve enerji tasarrufu sağlamak üzere saatlerin değiştirilmesidir. A = ABD standart YSU. Mart'ın ikinci Pazar günü başlar, Kasım'ın ilk Pazar günü biter; U = Bilinmiyor; N = Yaz saati uygulaması yok}
\item{saat_dilimi_iana}{GeoNames web hizmeti tarafından belirlendiği şekliyle IANA saat dilimi}
}}
\usage{havalimanları}
\description{Havalimanları hakkında faydalı kılavuz veriler.}
\seealso{\code{\link[nycflights13]{airports}}}
\keyword{datasets}
