\docType{data}
\name{havayolları}
\alias{havayolları}
\title{Havayolu adları}
\format{16 satır ve 2 sütundan oluşan bir data.frame
\describe{
\item{havayolu_kodu}{Havayolu şirketinin 2 harfli kısaltması}
\item{ad}{Havayolu şirketinin tam adı}
}}
\usage{havayolları}
\description{Havayolu adlarını taşıyıcı kodlarından arayın.}
\seealso{\code{\link[nycflights13]{airlines}}}
\keyword{datasets}
