\docType{data}
\name{kişiler}
\alias{kişiler}
\title{Kişiler tablosu}
\format{19.617 satır ve 26 sütundan oluşan bir data.frame
\describe{
\item{oyuncu}{Özgün oyuncu kodu}
\item{doğum_yılı}{Oyuncunun doğduğu yıl}
\item{doğum_ayı}{Oyuncunun doğduğu ay}
\item{doğum_günü}{Oyuncunun doğduğu gün}
\item{doğduğu_ülke}{Oyuncunun doğduğu ülke}
\item{doğduğu_eyalet}{Oyuncunun doğduğu eyalet}
\item{doğduğu_şehir}{Oyuncunun doğduğu şehir}
\item{ölüm_yılı}{Oyuncunun öldüğü yıl}
\item{ölüm_ayı}{Oyuncunun öldüğü ay}
\item{ölüm_günü}{Oyuncunun öldüğü gün}
\item{öldüğü_ülke}{Oyuncunun öldüğü ülke}
\item{öldüğü_eyalet}{Oyuncunun öldüğü eyalet}
\item{öldüğü_şehir}{Oyuncunun öldüğü şehir}
\item{adı}{Oyuncunun adı}
\item{soyadı}{Oyuncunun soyadı}
\item{tam_adı}{Oyuncunun adı ve göbek adı}
\item{libre}{Oyuncunun libre cinsinden vücut ağırlığı}
\item{boy}{Oyuncunun fit ve inç cinsinden boy uzunluğu}
\item{vuruş_eli}{Oyuncunun vuruş yaparken kullandığı el}
\item{atış_eli}{Oyuncunun atış yaparken kullandığı el}
\item{ilk_oyun_tarihi}{Birinci ligdeki ilk oyununun tarihi}
\item{son_oyun_tarihi}{Birinci ligdeki son oyununun tarihi (devam ediyorsa NA)}
\item{retro_no}{Retrosheet adresinde kullanılan oyuncu kodu \link{retrosheet.org}}
\item{bbref_no}{Baseball Reference adresinde kullanılan oyuncu kodu \link{baseball-reference.com}}
\item{ölüm_tarihi}{Oyuncunun yyyy-aa-gg formatında ölüm tarihi}
\item{doğum_tarihi}{Oyuncunun yyyy-aa-gg formatında doğum tarihi}
}}
\usage{kişiler}
\description{Oyuncu isimleri, doğum tarihleri ve biyografik bilgiler. Bu dosya, \code{\link{vuruşlar}}, \code{\link{atışlar}} vd. içerisindeki oyuncular hakkında \code{oyuncu} kodları yardımıyla daha fazla bilgi edinmek için kullanılabilir.}
\seealso{\code{\link[Lahman]{People}}}
\keyword{datasets}
