\docType{data}
\name{maaşlar}
\alias{maaşlar}
\title{Maaşlar tablosu}
\format{26.428 satır ve 5 sütundan oluşan bir data.frame
\describe{
\item{yıl}{Yıl}
\item{takım}{Takım kodu (faktör)}
\item{lig}{Lig kodu (faktör)}
\item{oyuncu}{Oyuncu kodu}
\item{maaş}{Dolar cinsinden maaş}
}}
\usage{maaşlar}
\description{Oyuncu maaşları verileri.}
\seealso{\code{\link[Lahman]{Salaries}}}
\keyword{datasets}
