\docType{data}
\name{mt_arabalar}
\alias{mt_arabalar}
\title{Motor Trend'in Araba Yol Testleri}
\format{32 gözlem 11 değişkenden oluşan bir data.frame
\describe{
\item{mil}{Galon başına katedilen mil}
\item{silindir}{Silindir sayısı}
\item{hacim}{Motor hacmi}
\item{beygir}{Brüt beygir gücü}
\item{aks}{Aks (diferansiyel veya son dişli) oranı}
\item{ağırlık}{Ağırlık (1000 libre)}
\item{hızlanma}{Çeyrek mil katetme süresi}
\item{tip}{Motor tipi (0 = V tipi, 1 = sıralı)}
\item{şanzıman}{Şanzıman türü (0 = otomatik, 1 = manuel)}
\item{vites}{İleri vites sayısı}
\item{karbüratör}{Karbüratör sayısı}
}}
\usage{mt_arabalar}
\description{Veriler, 1974 Motor Trend US dergisinden alınmıştır ve 32 otomobil (1973–74 modelleri) için yakıt tüketimini ve otomobil tasarımı ve performansının 10 özelliğini içermektedir.}
\seealso{\code{\link[datasets]{mtcars}}}
\keyword{datasets}
