\docType{data}
\name{ödüller}
\alias{ödüller}
\title{Takım koçları ödül tablosu}
\format{179 satır ve 6 sütundan oluşan bir data.frame
\describe{
\item{oyuncu}{Özgün oyuncu (koç) kodu}
\item{ödül}{Kazanılan ödülün adı}
\item{yıl}{Yıl}
\item{lig}{Lig kodu (faktör)}
\item{berabere}{Berabere kalma durumu. Ödülü birden çok kişi birlikte kazandıysa \code{E}, aksi taktirde \code{H}}
\item{notlar}{Ödülle ilgili notlar}
}}
\usage{ödüller}
\description{Takım koçlarının kazandığı ödüller}
\seealso{\code{\link[Lahman]{AwardsManagers}}}
\keyword{datasets}
