\docType{data}
\name{old_faithful}
\alias{old_faithful}
\title{Old Faithful gayzer verisi}
\format{272 satır ve 2 sütundan oluşan bir data.frame
\describe{
\item{püskürme_süresi}{Püskürme süresi (dakika)}
\item{bekleme_süresi}{Bir sonraki püskürme için bekleme süresi (dakika)}
}}
\usage{old_faithful}
\description{ABD, Wyoming, Yellowstone Ulusal Parkı'ndaki Old Faithful gayzeri için püskürme süreleri ve püskürmeler arası bekleme süreleri.}
\seealso{\code{\link[datasets]{faithful}}}
\keyword{datasets}
