\docType{data}
\name{savunma}
\alias{savunma}
\title{Savunma istatistikleri tablosu}
\format{140.921 satır ve 18 sütundan oluşan bir data.frame
\describe{
\item{oyuncu}{Özgün oyuncu kodu}
\item{yıl}{Yıl}
\item{takım_sırası}{Oyuncunun bir sezon içinde takımlar arası yaptığı geçiş sayısı}
\item{takım}{Takım kodu (faktör)}
\item{lig}{Lig kodu (faktör) (AA, AL, FL, NL, PL, UA)}
\item{mevki}{Savunma oyuncusunun mevkisi}
\item{oyun}{Oyun sayısı}
\item{başlama}{Oyuna başlama sayısı}
\item{oyun_süresi}{Sahada kaldığı süre}
\item{putout}{Oyundan çıkarma}
\item{asist}{Asist sayısı}
\item{hata}{Oyuncu tarafından yapılan kural hatası sayısı}
\item{double_play}{Savunma oyuncularının hücum oyuncularını art arda dışarıda bırakması}
\item{kaçan_top}{Tutucunun başarısız olduğu, tutamadığı top sayısı}
\item{karavana}{Atıcının topu doğru bölgeye fırlatamadığı atış sayısı}
\item{kale_çalma}{Tutucuların rakipten kale çalma sayısı}
\item{yakalanma}{Kale çalerken yakalanma sayısı}
\item{bölge_notu}{Savunmacının bölgelere göre kaç koşuyu engellediğini gösteren genel bir başarı ölçüsü}
}}
\usage{savunma}
\description{Savunma istatistikleri}
\seealso{\code{\link[Lahman]{Fielding}}}
\keyword{datasets}
