\docType{data}
\name{süsen}
\alias{süsen}
\title{Edgar Anderson'ın Süsen (Iris) çiçeği türleri verileri}
\format{150 gözlem ve 5 degişken içeren bir data.frame
\describe{
\item{çanak_uzunluğu}{Çanak yapraklarının uzunluğu}
\item{çanak_genişliği}{Çanak yapraklarının genişliği}
\item{taç_uzunluğu}{Taç yapraklarının uzunluğu}
\item{taç_genişliği}{Taç yapraklarının genişliği}
\item{tür}{Süsen çiçeklerinin türleri (Süsen setosa, S. versicolor, S. virginica)}
}}
\usage{süsen}
\description{Ünlü Süsen (Iris) verisi (Edgar Anderson ve Ronald Fisher) üç süsen türünün her biri için 50'şer çiçeğin, taç ve çanak yaprağı uzunluk ve genişlik ölçülerini santimetre cinsinde içerir. Bu türler Süsen setosa, S. versicolor, S. virginica olarak isimlendirilir.}
\seealso{\code{\link[datasets]{iris}}}
\keyword{datasets}
