\docType{data}
\name{tablo1}
\alias{tablo1}
\title{Dünya Sağlık Örgütü tüberküloz kayıtları (ilk varyant)}
\format{6 satır ve 4 sütundan oluşan bir data.frame
\describe{
\item{ülke}{Ülke adı}
\item{yıl}{Kayıtlı değerlerin gözlendiği yıl}
\item{vaka_sayısı}{Kayıtlı vaka sayısı}
\item{nüfus}{Ülke nüfusu}
}}
\usage{tablo1}
\description{Afganistan, Brezilya ve Çin'de, 1999 ve 2000 yıllarına ait kayıtlı vakaları içermektedir.}
\seealso{\code{\link[tidyr]{table1}}}
\keyword{datasets}
