\docType{data}
\name{tablo2}
\alias{tablo2}
\title{Dünya Sağlık Örgütü tüberküloz kayıtları (ikinci varyant)}
\format{12 satır ve 4 sütundan oluşan bir tibble
\describe{
\item{ülke}{Ülke adı}
\item{yıl}{Kayıtlı değerlerin gözlendiği yıl}
\item{kategori}{Gözlem değerinin ait olduğu kategori}
\item{sayım}{Kategoriler için gözlenen değer}
}}
\usage{tablo2}
\description{Afganistan, Brezilya ve Çin'de, 1999 ve 2000 yıllarına ait kayıtlı vakaları içermektedir.}
\seealso{\code{\link[tidyr]{table2}}}
\keyword{datasets}
