\docType{data}
\name{tablo3}
\alias{tablo3}
\title{Dünya Sağlık Örgütü tüberküloz kayıtları (üçüncü varyant)}
\format{6 satır ve 3 sütundan oluşan bir tibble
\describe{
\item{ülke}{Ülke adı}
\item{yıl}{Kayıtlı değerlerin gözlendiği yıl}
\item{oran}{Tüberküloz vakalarının toplam nüfusa oranı}
}}
\usage{tablo3}
\description{Afganistan, Brezilya ve Çin'de, 1999 ve 2000 yıllarına ait kayıtlı vakaları içermektedir.}
\seealso{\code{\link[tidyr]{table3}}}
\keyword{datasets}
