\docType{data}
\name{tablo4a}
\alias{tablo4a}
\title{Dünya Sağlık Örgütü tüberküloz kayıtları (dördüncü varyant, a)}
\format{3 satır ve 3 sütundan oluşan bir tibble
\describe{
\item{ülke}{Ülke adı}
\item{1999}{1999'da kaydedilen vaka sayısı}
\item{2000}{2000'de kaydedilen vaka sayısı}
}}
\usage{tablo4a}
\description{Afganistan, Brezilya ve Çin'de, 1999 ve 2000 yıllarına ait kayıtlı vakaları içermektedir.}
\seealso{\code{\link[tidyr]{table4a}}}
\keyword{datasets}
