\docType{data}
\name{tablo4b}
\alias{tablo4b}
\title{Dünya Sağlık Örgütü tüberküloz kayıtları (dördüncü varyant, b)}
\format{3 satır ve 3 sütundan oluşan bir tibble
\describe{
\item{ülke}{Ülke adı}
\item{1999}{1999 yılındaki nüfus}
\item{2000}{2000 yılındaki nüfus}
}}
\usage{tablo4b}
\description{Afganistan, Brezilya ve Çin'in 1999 ve 2000 yıllarına ait nüfus sayılarını içerir.}
\seealso{\code{\link[tidyr]{table4b}}}
\keyword{datasets}
