\docType{data}
\name{tablo5}
\alias{tablo5}
\title{Dünya Sağlık Örgütü tüberküloz kayıtları (beşinci varyant)}
\format{6 satır ve 4 sütundan oluşan bir tibble
\describe{
\item{ülke}{Ülke adı}
\item{yüzyıl}{Değerlerin kaydedildiği yüzyıl}
\item{yıl}{Kayıtlı değerlerin gözlendiği yıl}
\item{oran}{Tüberküloz vakalarının toplam nüfusa oranı}
}}
\usage{tablo5}
\description{Afganistan, Brezilya ve Çin'de, 1999 ve 2000 yıllarına ait kayıtlı vakaları içermektedir.}
\seealso{\code{\link[tidyr]{table5}}}
\keyword{datasets}
