\docType{data}
\name{taşıtlar}
\alias{taşıtlar}
\title{Taşıtların yakıt ekonomileri}
\format{33442 satır ve 12 sütundan oluşan bir tibble
\describe{
\item{no}{Özgün EPA numarası}
\item{üretici}{Üretici}
\item{model}{Model adı}
\item{yıl}{Model yılı}
\item{sınıf}{EPA'ya göre taşıt boyutu sınıfları. http://fueleconomy.gov/feg/ws/wsData.shtml#VClass}
\item{şanzıman}{Şanzıman türü}
\item{çekiş}{Çekiş türü}
\item{silindir}{Silindir sayısı}
\item{motor}{Motor hacmi (litre)}
\item{yakıt}{Yakıt türü}
\item{otoyol}{Otoyolda mil başına galon tüketimi}
\item{şehir}{Şehir içi mil başına galon tüketimi}
}}
\usage{taşıtlar}
\description{ABD Çevre Koruma Ajansı (EPA), 1985-2015 yılları arası yakıt ekonomisi verileri. Bu veri kümesi, sadece seçili değişkenleri ve eksik veri içermeyen taşıtları içerir.}
\seealso{\code{\link[fueleconomy]{vehicles}}}
\keyword{datasets}
