\docType{data}
\name{uçaklar}
\alias{uçaklar}
\title{Uçak kılavuz verileri}
\format{3.322 satır ve 9 sütundan oluşan bir data.frame
\describe{
\item{kuyruk_numarası}{Uçak kuyruk numarası}
\item{yıl}{Üretim yılı}
\item{tip}{Uçak tipi}
\item{üretici}{Üretici firma}
\item{model}{Uçağın modeli}
\item{motor_sayısı}{Motor sayısı}
\item{koltuk_sayısı}{Koltuk sayısı}
\item{hız}{Mil bölü saat cinsinden ortalama seyir hızı}
\item{motor}{Motor tipi}
}}
\usage{uçaklar}
\description{FAA uçak sicilinde bulunan tüm uçak kuyruk numaraları için kılavuz verileri. American Airways (AA) ve Envoy Air (MQ) kuyruk numaraları yerine filo numaralarını bildirdikleri için eşleştirilemez.}
\seealso{\code{\link[nycflights13]{planes}}}
\keyword{datasets}
