% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veriler.R
\docType{package}
\name{veriler-package}
\alias{veriler}
\alias{veriler-package}
\title{veriler: ``R for Data Science'' kitabında kullanılan verilerin Türkçe çevirileri}
\description{
"R for Data Science" kitabında kullanılan verilerin Türkçe çevirilerini içeren bir paket. Paket içerisindeki dahili fonksiyonlar ve YAML yapıları, aynı verilerin İspanyolca çevirilerini içeren \href{https://github.com/cienciadedatos/datos/}{datos} paketinden alınmıştır.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/botan/veriler}
  \item Report bugs at \url{https://github.com/botan/veriler/issues}
}

}
\author{
\strong{Maintainer}: Botan Ağın \email{aginbotan@gmail.com} (\href{https://orcid.org/0000-0002-1288-9770}{ORCID})

}
\keyword{internal}
