\docType{data}
\name{yaygın}
\alias{yaygın}
\title{Yaygın taşıt modelleri}
\format{347 satır ve 4 sütundan oluşan bir tibble
\describe{
\item{marka}{Marka}
\item{model}{Model}
\item{FALSE}{Model için toplam adet}
\item{n_yıl}{Toplam model-yıl adedi}
}}
\usage{yaygın}
\description{Bu modeller en az 10 yıllık veriye sahiptir. Bu veri kümesi, \code{\link{taşıtlar}} ile birleştirilerek (left-join), gözlem sayısını daha elverişli bir seviyeye çekmek için tasarlanmıştır.}
\seealso{\code{\link[fueleconomy]{common}}}
\keyword{datasets}
