\docType{data}
\name{havayolları}
\alias{havayolları}
\title{Havayolu adları}
\format{16 gözlem 2 değişkenden oluşan bir data.frame
\describe{
\item{taşıyıcı_kodu}{Taşıyı firmanın 2 harfli kısaltması}
\item{ad}{Firmanın tam adı}
}}
\usage{havayolları}
\description{Havayolu adlarını taşıyıcı kodlarından arayın.}
\seealso{\code{\link[nycflights13]{airlines}}}
\keyword{datasets}
